Reduce fuzz runs when generating CI coverage report
This is to reduce the time of generating the coverage report, in which we are not interested in
running lots of scenarios, but rather just making sure the tests exercise all the system's code.

This also turns off invariant tests. This codebase doesn't yet have these, but they really should
not be considered in coverage reports anyway.