Implement a wrapper for the LST
The wrapper converts the LST from a rebasing token with a dynamic balance to a
vault-like token where balances are not dynamic, but the exchange rate for the
underlying asset changes. The wrapper has an owner, and the owner can specify
and update the delegatee for the wrapper. This means all holders of any instance
of the wrapped LST will assign their voting weight to the same delegatee. The primary
use case for the wrapper is to allow usage of the LST as collateral in protocols.
Different instances of the wrapper, with different delegatees, can be deployed for use
as collateral in different protocols.