When adding to a delegated balance checkpoint, calculate amount from shares issued
When we add to a holder's delegated balance checkpoint, we shouldn't simply add the
amount directly. Instead, we should calculate the amount to be added by using the
underlying shares that were added/transferred in the process. If we don't do it this
way, we can find ourselves in situations where the holder's balance checkpoint is
actually greater than the calculated balance, since the latter is subject to truncation
that comes with converting shares to stake tokens.