Implement fixes to manage truncation issues and ensure solvency
When staking, unstaking, and transferring, an amount denominated in (st)UNI is provided
to the function being called. This amount is then used to calculate a number of shares
that must be moved. Due to truncation, after shares are moved, the *actual* amount that
changes can be slightly less than the amount requested by the caller. If handled naively,
it can lead to the LST, or to the underlying UniStaker deposit managing UNI for a given
delegatee, to become insolvent. To deal with this, we now use the provided amount to calculate
shares, then recalculate the *actual balance change* as a result of moving the shares. We then
operate on this second amount that has already been subject to the truncation, and thus ensure
that errors favor solvency of the LST in general, and each deposit in particular.