Update the mechanics used to move tokens between deposits during transfers
We update the internals of transfer to change the way underlying tokens are moved
between Staker dpeosits. Instead of naively withdrawing all the senders tokens,
moving the ones being sent, and consolidating the sender's tokens back to the sender's
deposit, we instead withdraw only the tokens needed to complete the transfer. Further,
we source the tokens with the same algorithm used to source them when a user unstakes.
That is, we first withdraw tokens from their undelegated balance, and only afterwards
withdraw however many tokens are still needed from their delegated balance.