Fix oversight wherein global parameters were not decremented upon unstake (#45)
* Fix oversite wherein global parameters were not decremented upon unstake

* Update gas scenarios to use the real withdraw gate for accurate unstake gas useage

* invariant unstake fix

* Handle the case when a user's balance is above their checkpoint due to truncation

A user's balance can change without the user taking an action due to actions of other
users. This occurs when changes in the total supply and total shares result in the
user's balance changing by tiny amounts due to truncation. Because of this, a user's
delegated balance checkpoint may actually be higher than their current balance.

This commit accomodates this by checking for that case in the instances when the user's
delegated balance checkpoint is used to calculate tokens that should be moved from the
user's chosen UniStaker deposit. In those instances that the delegated balance check-
point is above the user's live calculated balance, we simply use the live balance
instead. This means a few wei may be left behind in the deposit the user had specified.
This is fine, as both the old deposit and the new deposit remain solvent after the user
action has completed.

This commit also includes fuzz  tests for transfer, update deposit, and unstake which
are specifically tuned to test the case when this change due to truncation occurs.

---------

Co-authored-by: wildmolasses <changes@gmail.com>