Implement tests and a fix for transfer cases where truncation favors the sender (#62)
* Fix oversite wherein global parameters were not decremented upon unstake

* Implement tests & fix for transfer cases where truncation favors sender

In certain cases, it is possible for truncation to favor the sender—that is, the sender's
balance will decrease by an amount less than the receiver's will increase. This commit
adds explicit unit tests with values initially discovered by the fuzzer, to ensure we're
always testing for this case. It also implements a fix, whereby we shave shares from the
sender in such cases to ensure the underlying Staker deposits remain solvent.

* run gas report again

---------

Co-authored-by: wildmolasses <changes@gmail.com>