Reduce share scale factor to 1e10 and fix overly constrained tests
After reducing the share scale factor to 1e10, we see a number of failing tests.
In most cases, that is because these tests are overly conservative in their requiremnts
for testing the correct result. We modify the tests to allow for acceptable error rates
due to truncation to get the test suite passing again.

Reducing the scale factor may be useful in allowing us to pack various values into slots
by using a smaller integer size for shares.