Avoid tracking delegated balance checkpoint if the holder uses default deposit
If an LST token holder is using the default deposit, then there is no need to track
their delegated balance checkpoint. We know by definition that their full balance
should be in the default deposit at all times. This PR avoids tracking that checkpoint
to save gas for those users.