Minimize storage reads/writes by performing calculations in memory before writing to storage
For all operations that read and write to storage, first move the necessary data into memory,
perform all calculations and updates there, then write final values out to storage before
returning. Due to a stack too deep error in the transfer method, we also had to switch to
make some modifications to the transfer method to allow this change to work there.

Co-authored-by: wildmolasses <changes@gmail.com>