Add Default Delegatee Guardian & control mechanic for updating the default delegatee
This commit introduces a default delegatee guardian, a special role on the LST that has the
right to update the default delegatee. If the guardian has never exercised its right, then the
LST owner has the right to update the default delegatee, or the guardian itself. However, once
the guardian takes any action—either updating the defualt delegatee or itself—then it is
henceforth in control of these parameters. The LST owner is no longer able to update the default
delegatee or change the guardian.

When the LST is deployed, it is expected the owner will be the Tally DAO and the guardian will be
the Uniswap DAO itself. Should the Uniswap DAO choose to exercise its right to set the default delegatee
on the LST, it will thereafter be irrevocably in control of this parameter, and Tally DAO would be power-
less to change it.