Change how truncation is handled
Rather than comparing the live balance to the balance checkpoint at the time of a user action,
we now store the lesser as the balance checkpoint when an action occurs. We also round share
calculations up or down during a user action when calculating the shares to move or modify. In
combination, these two changes clean up our truncation mitigations, allowing us to remove the
need to shave shares during transfers. It also allows us to be mathematically certain that the
error introduced by truncation by any given operation will always be at-most 1 wei, as long as
raw total supply is less than raw shares. This is essentially guarenteed to be the case, since
raw shares have a large scale factor applied to them.