Implement and test a fixed balance counterpart to the core rebsaing LST contract
This commit introduces a new ERC20 token which is meant to be a counterpart to the
rebasing LST contract. Whereas the latter calculates the live balance of the holder,
including of rewards distributed to them, such that there balances always aligns 1:1
with the underlying governance tokens the hold, the new fixed balance LST exposes scaled
down shares as the users balance. This balance does not change, even as rewards are
distributed. Instead, the amount of underlying governance tokens the user has a claim to
increases as rewards are distributed.

The new fixed balances LST is a coupled counterpart that is deployed alongside the rebasing
LST. It uses priveleged helper methods on the LST to hook into the rebasing LSTs accounting
system, using alias addresses to perform operations on behalf the user who desires the
fixed LST instead. Whereas the rebasing LST provides the best UX for a holder, the fixed LST
provides easier integration for other protocols and for infrastructure providers like custody
solutions.

Users may stake or unstake directly in either version of the LST. They may delegate directly in
either version of the LST. And they may convert freely between one version and the other.