Refactor LST so that events are not emitted with Fixed LST alias addresses (#153)
The usage of alias addresses inside the rebasing LST to enable the fixed LST functionality is
an implementation detail. To whatever extent possible, we want to avoid leaking this implementation
detail outside the internals of the system itself. To this end, this commit refactors event emission
inside the rebasing LST to make it possible to avoid emitting events when the address involved would
be a fixed lst address alias.

In addition, we *add* events to the fixed lst helper methods inside the rebasing lst in order to make
the system more legible offchain. In general, we conceptualize the fixed LST as one would a wrapper
contract interacting with the rebasing LST. As such, events are included to mirror this external
conceptualization.