Update the way fixed address aliases are calculated
This change prevents a user from calculating the "inverse offset" of their address,
and therefore prevents a holder of the Fixed LST from sending tokens to this inverse
address. If a user were to do so, these tokens would exit the Fixed LST and return to
the Rebasing LST, but without the shares being properly removed from the total supply
of the Fixed LST. As a result, the Fixed LST total supply could reflect an unrealistic
value.