Remove tips for caller of override methods
This PR rips out the issuing of dilutive shares to the caller of the override family
of methods. These tips were intended as an incentive for callers of the override methods,
however their use breaks the internal accounting of stake tokens across the deposits
owned and managed by the LST. There are weak incentives for all LST holders to call the
override methods (to maximize their yields), and strong incentives for the owner of the
LST to call these methods to ensure the LST operates at maximum yield. We will rely on
these incentives instead.

Furthermore, it would be possible to build and deploy an incentivization scheme as an
external contract at a later date. This could be achieved by directing (a portion) of
LST protocol fees to a contract that rewards pass-through callers of the override methods.